move `getKafkaWorkflow()` out of `commands/docker` for easier testability
